Initialize Bazel. Create super-basic container image for Genesis (#4)
This PR is part of the Genesis project. [This is the implementation plan
for Genesis][1].

## Changes

### Initialize Bazel

This PR initializes this repository's support for Bazel.

* We use [Bazelisk][2] to install a consistent version of Bazel.
    * The Bazel version is set in the `.bazeliskrc` file.
* `.bazelrc` sets global setting such as `common --enable_bzlmod`, which
opts us into the [Bzlmod][3].

### Genesis container image

The image has the following characteristics.

* Contains the terraform binary.
* Contains empty terraform code.

We can make basic structure assertions using `bazel test
//abacus/genesis:source_code_layer_test`, which leverages [the container
structure test rules for Bazel][5].

## Functional results

The engineer can run the following commands locally.

* `bazel build //abacus/genesis:source_code_layer`
* `bazel test //abacus/genesis:source_code_layer_test`
* `bazel run //abacus/genesis:genesis_image`

## Manual operations

@josalvatorre had to manually upload a base image to our public ECR
repository so that Bazel could consume it as a dependency. [The plan is
to automate this after solving the chicken-and-egg problem described
here][4].

## Why not add a `bazel test //...` GitHub Action?

I tried that in
https://github.com/josalvatorre/monorepo-alpha/pull/4/commits/c9248e4030a1119ee8737c0f3a8abee9e98b5a58,
but GitHub Actions workers use `x86`, so they failed to build.

I could work around this, but I'd rather set this up on AWS after
Genesis launches. AWS has `arm64` machines, so we wouldn't have to do
something super-hacky.

[1]:
https://github.com/josalvatorre/monorepo-alpha/blob/725c385bfe8e529dbb50980083544ffa2451986d/abacus/genesis/README.md?plain=1#L99
[2]: https://github.com/bazelbuild/bazelisk
[3]: https://bazel.build/external/migration
[4]:
https://github.com/josalvatorre/monorepo-alpha/blob/725c385bfe8e529dbb50980083544ffa2451986d/abacus/genesis/README.md?plain=1#L103-L107
[5]: https://github.com/GoogleContainerTools/container-structure-test