Manually set up Terraform auth (#8)
## Changes

To get automatic Terraform deployments up and running, we must first set
up auth so that Terraform Cloud can access our AWS organization. We want
to do this via an OpenID provider so we don't rely on hard-coded
credentials. I'm following the advice in [this example produced by
HashiCorp][1].

The code in this PR was executed manually and locally on my machine
using the Terraform CLI. In my next PR, I will activate cloud-based
deployments.

## Manual deployment process

Do the following in *the same terminal session*.
1. `cd abacus/genesis/`
2. `terraform login`
3. Set your AWS credentials for the `abacus-org` account.

[1]:
https://github.com/hashicorp/terraform-dynamic-credentials-setup-examples/tree/5308cd970c0832f2180d7eb1e645dea33c4e344c/aws