Use default project again and add TFE_TOKEN as a variable (#16)
## Changes

Based on my research, the likely cause of our missing project problem is
that we haven't provided a `TFE_TOKEN`. This is weird, given that we're
running these on Terraform Cloud. It's also weird that Terraform doesn't
have a simple way to set up auto-rotating tokens.

This PR adds `TFE_TOKEN` as a required variable. The operator will need
to mint short-lived tokens for every deployment 😭 Thankfully, Terraform
deployments shouldn't be as common as others, so I hope I don't need to
automate this.

## Resources

*
https://registry.terraform.io/providers/hashicorp/tfe/latest/docs#authentication
*
https://developer.hashicorp.com/terraform/cloud-docs/users-teams-organizations/api-tokens
* https://github.com/hashicorp/terraform-provider-tfe/issues/882
*
https://support.hashicorp.com/hc/en-us/articles/30556582962707-HCP-Terraform-Resolving-Resource-Not-Found-Error-When-Using-TFE-Provider-in-Terraform-Cloud

### Misleading rotation snippet

You might find snippets in the resources above mentioning rotation.
However, I found these to be misleading. At the time of writing, there's
no way to create an auto-rotating token for Terraform Cloud. You could
*simulate* this by setting up an expiration lifecycle much slower than
your deployment rate. For example, if you deploy every few minutes and
set each token to expire every hour, then Terraform would be able to
replace the token. However, I don't deploy often enough for this to make
sense. I would need to set the expiration date to be a long time. It's
better to bite the bullet and have a convention around always minting
tokens per deployment.